Revised how commands are processed to be all at once Improved how multiple commands are processed by cutting on the command prompt Updated system message to direct llm to not try to predict command output Show waiting msg in console while llm is processing Added update and cycle detection scripts Updated packages

+++ src/shellCommand.ts
@@ -1,20 +1,20 @@
 import * as contextManager from "./contextManager.js";
 import * as inputMode from "./inputMode.js";
 import { InputMode } from "./inputMode.js";
-import * as promptBuilder from "./promptBuilder.js";
 import * as shellWrapper from "./shellWrapper.js";
 
 interface HandleShellCommandResponse {
   commandHandled: boolean;
-  processNextLine?: boolean;
   terminate?: boolean;
 }
 
 export async function handleCommand(
-  line: string,
-  consoleInputLines: string[],
+  input: string,
 ): Promise<HandleShellCommandResponse> {
-  const cmdParams = line.trim().split(" ");
+  const cmdParams = input.split(" ");
+  const response: HandleShellCommandResponse = {
+    commandHandled: true,
+  };
 
   // Route user to context friendly edit commands that can read/write the entire file in one go
   if (["nano", "vi", "vim"].includes(cmdParams[0])) {
@@ -22,52 +22,35 @@ export async function handleCommand(
       `${cmdParams[0]} not supported. Use 'cat' to view a file and 'cat > filename << EOF' to write a file`,
     );
 
-    return {
-      commandHandled: true,
-    };
+    return response;
+  }
+
+  if (cmdParams[0] == "lynx" && cmdParams[1] != "--dump") {
+    contextManager.append(
+      `Interactive mode with lynx is not supported. Use --dump with lynx to view a website`,
+    );
+
+    return response;
   }
 
   if (cmdParams[0] == "exit") {
-    let terminate = false;
-
     if (inputMode.current == InputMode.LLM) {
       contextManager.append(
         "Use 'endsession' to end the session and clear the console log.",
       );
     } else if (inputMode.current == InputMode.Debug) {
       await shellWrapper.terminate();
-      terminate = true;
+      response.terminate = true;
     }
 
-    return {
-      commandHandled: true,
-      terminate,
-    };
+    return response;
   }
 
-  let allInput = line;
-  const promptPrefix = promptBuilder.getPromptPrefix();
-
-  while (consoleInputLines.length) {
-    const nextLine = consoleInputLines.shift() || "";
-    if (nextLine.startsWith(promptPrefix)) {
-      consoleInputLines.unshift(nextLine);
-      break;
-    } else {
-      contextManager.append(nextLine, "llm");
-    }
-
-    allInput += "\n" + nextLine;
-  }
-
-  const output = await shellWrapper.executeCommand(allInput);
+  const output = await shellWrapper.executeCommand(input);
 
   if (output) {
     contextManager.append(output);
   }
 
-  return {
-    commandHandled: true,
-    processNextLine: false,
-  };
+  return response;
 }
